game rule
This game is made by modifying the existing blackjack rules. In Blackjack, the game is played with a dealer, but in Jeongho Jack, there is no dealer. After all players are dealt two cards, it is up to each player to take an additional card. However, when there are two or more players with the highest number, the player with the smallest card wins. The player closest to 21 wins.